backend/requirements.txt
agent-framework[all]==1.0.0rc5
azure-ai-evaluation>=1.0.0
azure-identity>=1.19.0
azure-cosmos>=4.9.0
azure-ai-contentsafety>=1.0.0
azure-communication-sms>=1.1.0
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
websockets>=14.0
httpx>=0.28.0
python-dotenv>=1.0.1
pydantic>=2.10.0
opentelemetry-api>=1.29.0
opentelemetry-sdk>=1.29.0
opentelemetry-exporter-otlp>=1.29.0
opentelemetry-instrumentation-fastapi>=0.50b0
azure-monitor-opentelemetry-exporter>=1.0.0b33
twilio>=9.4.0
